Implement Copilot review feedback improvements
Critical fixes:
- ✅ Implement bandit feedback loop - now updates bandit statistics from user feedback
- ✅ Fix org_key comparison in feedback service to handle type consistency
- ✅ Remove unnecessary backwards compatibility check in codegen router
- ✅ Fix schema count fields to use int instead of float for semantic correctness
- ✅ Remove unreachable fallback code in Thompson Sampling algorithm

User experience improvements:
- ✅ Add user-facing error messages in FeedbackBar component
- ✅ Enhanced error handling in AdminFeedbackStats with specific failure messages
- ✅ Improve error display with clear UI indicators and actionable messages

Code quality:
- ✅ Add proper logging and error handling throughout
- ✅ Fix type consistency and validation issues
- ✅ Remove dead code and improve algorithm clarity"

+++ backend/schemas/ai_feedback.py
@@ -96,9 +96,9 @@ class ArmPerformance(BaseModel):
 
     model_config = ConfigDict(frozen=True)
 
-    successes: float = Field(..., description="Success count")
-    failures: float = Field(..., description="Failure count")
-    total_trials: float = Field(..., description="Total trials")
+    successes: int = Field(..., description="Success count")
+    failures: int = Field(..., description="Failure count")
+    total_trials: int = Field(..., description="Total trials")
     success_rate: float = Field(..., ge=0.0, le=1.0, description="Success rate")
     confidence: float = Field(..., ge=0.0, le=1.0, description="Confidence level")
